refactor: one link to the repository, in the shared ui
GithubRepoLink fixes the href, the tab and the rel, so no caller can get one of
the three wrong. The github mark it sits beside was two copies of one svg.
