+++ packages/ui/package.json
@@ -19,6 +19,7 @@
   "dependencies": {
     "@base-ui/react": "^1.6.0",
     "@fontsource-variable/source-code-pro": "^5.3.0",
+    "@repo/global-constants": "workspace:*",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "lucide-react": "catalog:",
